+++ example/lib/main.dart
@@ -74,6 +74,10 @@ class _MyHomePageState extends State<MyHomePage> {
       videoPath: videoPath,
       placementBuilder: _addPlacements,
       onTapEvent: _onTapEvent,
+      callType: EasyCallType.video,
+      callScheduleDuration: Duration(seconds: 7),
+      skipCallSettings: true,
+      ringtonePath: 'YOUR_AUDIO_URL / YOUR_AUDIO_ASSET',
     );
   }